requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1
torchvision==0.13.1
opencv-python==4.7.0.72
pudb==2019.2
imageio==2.9.0
pytorch-lightning==1.4.2
omegaconf==2.1.1
einops==0.3.0
kornia==0.6
torchmetrics==0.6.0
gradio==3.41.2
pillow==9.5.0
rembg==2.0.56
plotly==5.13.1
trimesh==3.23.5
yacs==0.1.8
dl-ext==1.3.4
git+https://github.com/openai/CLIP.git
loguru==0.7.2